new updates for products

+++ resources/views/branches/index.blade.php
@@ -30,9 +30,9 @@
                         <td>{{$branch->phone}}</td>
                         <td>{{$branch->email}}</td>
                         <td>
-                            <a href="#" onclick="openModal({'url':'###','modalId':'ajaxModal','method':'GET'})"
+                            <a href="#" onclick="openModal({'url':'{{route('branch.edit',[$branch->id])}}','modalId':'ajaxModal','method':'GET'})"
                                 class="std-btn-sm default">Edit</a>
-                            <a href="#" onclick="openModal({'url':'###','modalId':'ajaxModal','method':'DELETE'})"
+                            <a href="#" onclick="openModal({'url':'{{route('branch.confirm.delete',[$branch->id])}}','modalId':'ajaxModal','method':'DELETE'})"
                                 class="std-btn-sm default">Delete</a>
                         </td>
                     </tr>
